Replace EditorPopover (mac)
